requirements.txt
# Copyright IBM Corp. 2025
flask
flask_cors
python-dotenv
langchain
gunicorn
langchain-ibm